+++ src/vscodeUI.ts
@@ -1,13 +1,35 @@
 import * as fs from "fs";
+import * as os from "os";
 import * as vscode from "vscode";
 
 export class VSCodeUI {
-    public static runInTerminal(command: string, addNewLine: boolean = true, terminal: string = "Maven"): void {
-        if (this.terminals[terminal] === undefined) {
-            this.terminals[terminal] = vscode.window.createTerminal(terminal);
+    public static runInTerminal(command: string, options?: ITerminalOptions): void {
+        const defaultOptions: ITerminalOptions = { addNewLine: true, name: "Maven" };
+        const { addNewLine, name, cwd } = Object.assign(defaultOptions, options);
+        if (this.terminals[name] === undefined) {
+            this.terminals[name] = vscode.window.createTerminal({ name });
         }
-        this.terminals[terminal].show();
-        this.terminals[terminal].sendText(command, addNewLine);
+        this.terminals[name].show();
+        if (cwd) {
+            this.terminals[name].sendText(this.getCDCommand(cwd), true);
+        }
+        this.terminals[name].sendText(command, addNewLine);
+    }
+
+    public static getCDCommand(cwd: string): string {
+        if (os.platform() === "win32") {
+            const windowsShell = vscode.workspace.getConfiguration("terminal").get<string>("integrated.shell.windows")
+                .toLowerCase();
+            if (windowsShell && windowsShell.indexOf("bash.exe") > -1 && windowsShell.indexOf("git") > -1) {
+                return `cd "${cwd.replace(/\\+$/, "")}"`; // Git Bash: remove trailing '\'
+            } else if (windowsShell && windowsShell.indexOf("powershell.exe") > -1) {
+                return `cd "${cwd}"`; // PowerShell
+            } else if (windowsShell && windowsShell.indexOf("cmd.exe") > -1) {
+                return `cd /d "${cwd}"`; // CMD
+            }
+        } else {
+            return `cd "${cwd}"`;
+        }
     }
 
     public static onDidCloseTerminal(closedTerminal: vscode.Terminal): void {
@@ -35,4 +57,10 @@ export class VSCodeUI {
     }
 
     private static terminals: { [id: string]: vscode.Terminal } = {};
+}
+
+interface ITerminalOptions {
+    addNewLine?: boolean;
+    name?: string;
+    cwd?: string;
 }